+++ packages/contracts/src/contracts/interfaces/IRewardRedemptionFactory.sol
@@ -2,11 +2,10 @@
 pragma solidity 0.8.20;
 
 interface IRewardRedemptionFactory {
-   
-
- event RewardRedemptionCreated( address rewardRedemption, 
+    event RewardRedemptionCreated(
+        address rewardRedemption,
         bytes32 indexed appId,
         string name,
-        uint8 tokensRequired
-    ) ;
+        uint256 tokensRequired
+    );
 }
